scraping/requirements.txt
Package            Version
------------------ ---------
beautifulsoup4     4.12.3
certifi            2025.1.31
charset-normalizer 3.4.1
idna               3.10
pip                24.3.1
requests           2.32.3
soupsieve          2.6
urllib3            2.3.0
pillow             11.1.0
